Add the brave full version

+++ src/music_player/note/MelodyNote.h
@@ -8,15 +8,15 @@ class MelodyNote
 {
 private:
     Note _note;
-    uint32_t _duration;
-    uint32_t _duty;
-    uint32_t _pause;
+    uint16_t _duration;
+    uint16_t _duty;
+    uint16_t _pause;
 public:
-    MelodyNote(Note note, uint32_t duration, uint32_t duty, uint32_t pause) : 
+    MelodyNote(Note note, uint16_t duration, uint16_t duty, uint16_t pause) : 
         _note(note), _duration(duration), _duty(duty), _pause(pause) {};
 
     Note getNote() const {return _note;}
-    uint32_t getDuration() const {return _duration;}
-    uint32_t getDuty() const {return _duty;}
-    uint32_t getPause() const {return _pause;}
+    uint16_t getDuration() const {return _duration;}
+    uint16_t getDuty() const {return _duty;}
+    uint16_t getPause() const {return _pause;}
 };
